fix: reorganize structure and fix deploy method
