+++ functions/notify_token_service/line_notify_tokens.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass, field
 from typing import ClassVar, List
 
+from db import connect_with_connector, get_db_name, get_schema_name
 from sqlalchemy import MetaData, Table, insert
-
-from shared.db.connect import connect_with_connector, get_db_name, get_schema_name
 
 # 配置 SQLAlchemy 引擎
 engine = connect_with_connector()
