+++ functions/weekly_game_notify/games.py
