+++ functions/weekly_game_notify/gcloud.py
